add class manager with EntityManager

+++ src/main/java/ru/hiber/entity/Product.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Table(name = "product")
 public class Product {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private Double price;
 
@@ -26,7 +26,7 @@ public class Product {
         this.price = price;
     }
 
-    public Product(Integer id ,String name, Double price) {
+    public Product(Long id ,String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -44,7 +44,7 @@ public class Product {
         this.personList = personList;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -52,7 +52,7 @@ public class Product {
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
